add keep_registered option to hyperv

+++ builder/hyperv/common/step_clone_vm.go
@@ -31,6 +31,7 @@ type StepCloneVM struct {
 	SecureBootTemplate             string
 	EnableVirtualizationExtensions bool
 	MacAddress                     string
+	KeepRegistered                 bool
 }
 
 func (s *StepCloneVM) Run(ctx context.Context, state multistep.StateBag) multistep.StepAction {
@@ -147,6 +148,12 @@ func (s *StepCloneVM) Cleanup(state multistep.StateBag) {
 
 	driver := state.Get("driver").(Driver)
 	ui := state.Get("ui").(packer.Ui)
+
+	if s.KeepRegistered {
+		ui.Say("keep_registered set. Skipping unregister/deletion of VM.")
+		return
+	}
+
 	ui.Say("Unregistering and deleting virtual machine...")
 
 	err := driver.DeleteVirtualMachine(s.VMName)
